+++ client/lib/js/adapters.ts
@@ -914,20 +914,36 @@ namespace ABeamer {
 
       // @TODO: Find a way to sync video.
       // this code doesn't work on chrome.
+
       if (pos !== undefined) {
-        elMedia.currentTime = pos;
-        window.setTimeout(() => {
+        if (elMedia.readyState < 3) {
+          elMedia.addEventListener('loadeddata', () => {
+            _waitForMediaSync(elMedia, args, pos);
+          }, { once: true });
+          return;
+        }
+
+        elMedia.addEventListener('seeked', () => {
+          // #debug-start
+          if (_args.isVerbose) {
+            args.story.logFrmt(`video-sync: `, [
+              ['expected', pos],
+              ['actual', elMedia.currentTime],
+            ], LT_MSG);
+          }
+          // #debug-end
           onDone();
-        }, 1);
-        /*         elMedia.play().then(() => {
-                  elMedia.pause();
-                }); */
+        }, { once: true });
+        elMedia.currentTime = pos;
       } else {
         onDone();
       }
     }, {});
   }
 
+  /*         elMedia.play().then(() => {
+            elMedia.pause();
+          }); */
 
   export interface _WorkWaitForParams extends AnyParams {
     waitFor: WaitFor;