+++ client/lib/js/interpolator.ts
@@ -400,12 +400,12 @@ namespace ABeamer {
         }
 
         if (isDifferent) {
-          args.story.logFrmt(`action-update-warn: `,
-            [['id', elAdapter.getId(args)]
-              , ['prop', propName]
-              , ['expected', newValue + '']
-              , ['actual', actualNewValue + ''],
-            ], LT_WARN);
+          args.story.logFrmt(`action-update-warn: `, [
+            ['id', elAdapter.getId(args)],
+            ['prop', propName],
+            ['expected', newValue + ''],
+            ['actual', actualNewValue + ''],
+          ], LT_WARN);
         }
         log(propName, newValue);
       }