+++ client/lib/typings/release/abeamer-release.d.ts
@@ -698,6 +698,13 @@ declare namespace ABeamer {
      * @default #video
      */
     selector?: ElSelector;
+
+    /**
+     * If it's `false`, is disabled if `isServer = true`.
+     *
+     * @default true
+     */
+    serverRender: boolean;
   }
 
 
@@ -1027,6 +1034,12 @@ declare namespace ABeamer {
     waitMan?: WaitMan;
     /** If true, the internal code or plugin should dump information using `story.logFrmt`. */
     isVerbose?: boolean;
+    /**
+     * True if it's running a supporting server program for frame storage.
+     *
+     * #end-user @readonly
+     */
+    hasServer?: boolean;
     isTeleporting?: boolean;
     /** If true, perform type checks and other restriction checks. */
     isStrict?: boolean;