+++ gallery/animate-video-sync/js/main.ts
@@ -32,6 +32,7 @@ $(window).on("load", () => {
 
   story.addFlyover('video-sync', {
     selector: '#video',
+    serverRender: false,
   });
 
   story.render(story.bestPlaySpeed());