+++ server/server-agent-puppeteer.js
@@ -128,9 +128,10 @@ var ServerAgent;
                                     '--enable-fast-unload',
                                     '--disable-cloud-import',
                                     '--safebrowsing-disable-auto-update',
-                                    '--ignore-autoplay-restrictions',
                                     '--no-user-gesture-required',
                                     '--enable-font-antialiasing',
+                                    '--ignore-autoplay-restrictions',
+                                    '--autoplay-policy=no-user-gesture-required',
                                 ],
                             }
                                 : undefined)];